Split all table dialogs into separate Edit and ReadOnly components
The old pattern used a single dialog component with conditional
rendering based on isInEdit. The new pattern separates each dialog
into a focused EditDialog (validation, inline editing, error tracking)
and a lightweight ReadOnlyDialog (display only). This improves
maintainability and aligns with the pattern established by the
Synonyms/Symbol/FullName dialogs.

New dialog pairs:
- DatabaseStatus, GermlineTransmissionStatus, SecondaryIds,
MutationTypes, InheritanceModes, FunctionalImpacts,
NomenclatureEvents, RelatedNotes, ConditionRelations

Updated consumers:
- AllelesTable, VariantsTable, DiseaseAnnotationsTable (Edit + ReadOnly)
- GenesTable, ConstructsTable, AffectedGenomicModelTable,
PhenotypeAnnotationsTable, ComponentsDialog,
GenomicComponentsDialog (ReadOnly only)
