+++ src/main/cliapp/src/containers/constructsPage/GenomicComponentsDialog.jsx
@@ -8,7 +8,7 @@ import { Row } from 'primereact/row';
 import { EllipsisTableCell } from '../../components/EllipsisTableCell';
 import { evidenceTemplate } from '../../components/EvidenceComponent';
 import { Button } from 'primereact/button';
-import { RelatedNotesDialog } from '../../components/RelatedNotesDialog';
+import { RelatedNotesReadOnlyDialog } from '../../components/RelatedNotesReadOnlyDialog';
 import { getIdentifier } from '../../utils/utils';
 
 export const GenomicComponentsDialog = ({
@@ -181,12 +181,9 @@ export const GenomicComponentsDialog = ({
 					</DataTable>
 				</Dialog>
 			</div>
-			<RelatedNotesDialog
+			<RelatedNotesReadOnlyDialog
 				originalRelatedNotesData={relatedNotesData}
 				setOriginalRelatedNotesData={setRelatedNotesData}
-				errorMessagesMainRow={errorMessagesMainRow}
-				setErrorMessagesMainRow={setErrorMessagesMainRow}
-				noteTypeVocabularyTermSet="construct_component_note_type"
 			/>
 		</>
 	);
